backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
alembic==1.13.1
pymysql==1.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
